feat: Add server options
Added the ability to specify the protocol to serve with (http1 or http2), serve defaults to running with http/1.1 but to serve content with http/2, the option --protocol=h2 must be added. Any value other than h2 is not acceptible. Another option is --port which defaults to 3000. Also abstracted http2 server to its own file server/http2_server.js to clear things out