Add fingerprinted filenames to service worker
Since we don't know the fingerprint added to the filename of each file that went through revision gulp task, we need to add the new filenames dynamically to the service worker to cache them. This was done with static-module, by getting all the fingerprinted filenames from rev-manifest.json and adding them to the service worker as an array.